+++ RELEASE_NOTES.md
@@ -1,13 +1,13 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v1.1.46
+This is the offical release of vpinfe v1.1.47
 
 ### What's New
-None
+- Added a DOF event test panel to the DOF configuration panel. 
 
 ### Fixes
 None
 
 ### Notes
-Grab latest lib-python update. 
+None